Do not attempt to mutate dataset object
It turns out that `load_from_disk` returns a `Dataset` object that _looks_
like a list but isn't. So trying to mutate the dicts in place silently fails
to do anything at all.

Instead of trying to mutate the dataset in-place (which we'll copy into a list
anyway), just call a helper function that will deal with mangling the dict as
necessary while we enumerate dataset.